% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_dataset.R
\name{get_info_dataset}
\alias{get_info_dataset}
\title{get_info_dataset}
\usage{
get_info_dataset(
  url = "https://data.enedis.fr/api/explore/v2.1/catalog/",
  data_id = "production-electrique-par-filiere-a-la-maille-departement"
)
}
\arguments{
\item{url}{character: url de l'api explore/v2.1/catalog de Enedis}

\item{data_id}{character: Identifiant du dataset}
}
\value{
list avec 2 elements
nb_ligne: nombres de lignes dans la dataset
nom_colonnes: les noms de colonnes dans la dataset
}
\description{
Récuperer le nombre de lignes et les noms de colonnes d'une dataset opendata Enedis
}
\examples{
get_info_dataset(url="https://data.enedis.fr/api/explore/v2.1/catalog/",data_id="production-electrique-par-filiere-a-la-maille-departement")
}
