% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_row_dataset_api_enedis.R
\name{get_row_dataset_api_enedis}
\alias{get_row_dataset_api_enedis}
\title{get_row_dataset_api_enedis}
\usage{
get_row_dataset_api_enedis(
  url = "https://data.enedis.fr/api/records/1.0/search/",
  data_id = "production-electrique-par-filiere-a-la-maille-departement",
  debut_ligne = 0,
  nb_ligne = 10
)
}
\arguments{
\item{url}{character : url de l'api records/1.0/search de Enedis}

\item{data_id}{character: Identifiant du dataset}

\item{debut_ligne}{integer: Index du premier résultat renvoyé}

\item{nb_ligne}{integer: Nombre de lignes de résultat}
}
\value{
list avec une partie du dataset
}
\description{
Récuperer des lignes d'une dataset open data Enedis
}
\examples{
get_row_dataset_api_enedis(url="https://data.enedis.fr/api/records/1.0/search/",data_id="production-electrique-par-filiere-a-la-maille-departement",debut_ligne=0,nb_ligne=10)
}
